+++ tests/utils/test_get_value_from_or_update_nested_dict.py
@@ -1,5 +1,6 @@
 import pytest
 
+from alltheutils.exceptions import NestedDictExceptions
 from alltheutils.utils import get_value_from_or_update_nested_dict
 
 
@@ -17,7 +18,7 @@ def test_deep_retrieve():
 # TODO: differentiate between non-existent path and non-dict path
 def test_error_on_retrieving_nonexistent_path():
     data = {"you": {"me": "huh?"}}
-    with pytest.raises(KeyError):
+    with pytest.raises(NestedDictExceptions.ValueDoesNotExist):
         get_value_from_or_update_nested_dict(data, "you.nonexistent")
 
 
@@ -53,13 +54,14 @@ def test_retrieve_from_list_using_integer_key():
 
 def test_error_on_get_from_list_using_explicitly_string_integer_key():
     data = {"users": [{"name": "Alice"}, {"name": "Bob"}]}
-    with pytest.raises(TypeError):
+    with pytest.raises(NestedDictExceptions.ValueNotADict):
         get_value_from_or_update_nested_dict(data, "users.str>1.name")
+
 
 def test_error_on_get_from_list_using_implicitly_string_integer_key():
     data = {"users": [{"name": "Alice"}, {"name": "Bob"}]}
-    with pytest.raises(TypeError):
-        get_value_from_or_update_nested_dict(data, "users.str>1.name")
+    with pytest.raises(NestedDictExceptions.ValueNotADict):
+        get_value_from_or_update_nested_dict(data, "users.1.name")
 
 
 # ================================== Update ====================================
@@ -95,14 +97,13 @@ def test_replace_entire_dict_empty_address():
 
 def test_replace_entire_dict_with_non_dict():
     data = {"old": "data"}
-    with pytest.raises(TypeError):
+    with pytest.raises(NestedDictExceptions.NonDictReplacementValue):
         get_value_from_or_update_nested_dict(data, ".", "just a string")
 
 
-# TODO: differentiate between non-existent path and non-dict path
 def test_error_on_non_dict_path():
     data = {"you": {"me": "huh?"}}
-    with pytest.raises(TypeError):
+    with pytest.raises(NestedDictExceptions.ValueNotADict):
         get_value_from_or_update_nested_dict(data, "you.me.key", "fail")
 
 
@@ -126,7 +127,7 @@ def test_update_dict_with_implicitly_string_integer_key():
 
 def test_error_on_update_nested_dict_with_list_using_implicitly_string_integer_key():
     data = {"users": [{"name": "Alice"}, {"name": "Bob"}]}
-    with pytest.raises(TypeError):
+    with pytest.raises(NestedDictExceptions.ValueNotADict):
         get_value_from_or_update_nested_dict(data, "users.0.name", "Charlie")
 
 
@@ -168,7 +169,7 @@ def test_address_not_existing():
 
 def test_non_dict_middle_element():
     data = {"top": "string_value"}
-    with pytest.raises(TypeError):
+    with pytest.raises(NestedDictExceptions.ValueNotADict):
         get_value_from_or_update_nested_dict(data, "top.sub.value", "fail")