+++ docs/changelog.md
@@ -11,22 +11,48 @@ This software uses [Semantic Versioning v2.0.0](https://semver.org/spec/v2.0.0.h
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
-## 2.0.0 (Unreleased)
+## 3.0.0 (Unreleased)
+
+### Removed
+
+- Functions deprecated in v2.0.0
+
+## 2.0.0
+
+### Added
+
+- `alltheutils.base_exc.deprecated` decorator for deprecated base_exc function
+
+### Changed
+
+- Exceptions in `alltheutils.utils.get_value_from_or_update_nested_dict` to be custom and more descriptive
+- Parameter names of function `alltheutils.utils.if_none`
+
+### Deprecated
+
+- The following modules has been renamed:
+    - `alltheutils.base_exc` -> `alltheutils.base_exceptions`
+
+- The following functions has been renamed to be more readable (note that the parameters' names might be changed):
+    - `alltheutils.base_exc.c_exc` -> `alltheutils.base_exceptions.custom_exception`
+    - `alltheutils.base_exc.c_exc_str` -> `alltheutils.base_exceptions.custom_exception_str`
 
 ### Removed
 
 - Functions deprecated in v1.1.0
+- `alltheutils.exceptions.CDExceptions`
 
 ## 1.1.0
 
 ### Added
 
+- `alltheutils.utils.deprecated` decorator for deprecated utils function
 - `alltheutils.utils.get_value_from_or_update_nested_dict`
 - MIT License
 
 ### Deprecated
 
-- The following functions and renamed them to be more readable (note that the parameters' names might be changed):
+- The following functions has been renamed to be more readable (note that the parameters' names might be changed):
     - `alltheutils.utils.calc_hash` -> `calculate_sha256_hash`
     - `alltheutils.utils.cycle_2ls` -> `zip_extend`
     - `alltheutils.utils.dnn` -> `parent_dir_nth_times`